atc, fly: remove RawConfig/Errors logic
this was a holdover from when we used to store pipeline configs
wholesale in the database and had to worry about older,
no-longer-compatible pipeline configs in the db.

since then, we've switched to storing resource/job configs individually
and constructing the pipeline config in the handler, so it's no longer
possible for 'Errors' to be returned or for 'RawConfig' to be anything
but a re-marshalled 'Config'. so this just removes all the code for
handling that.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>